split out LCD routines
  reasonable demo LCD

	new file:   src/lcd_uses.cpp
	new file:   src/lcd_uses.h
	modified:   src/main.cpp

On branch dvlp-aa
