+++ src/lcd_uses.h
@@ -1,0 +1,7 @@
+/** lcd_uses.h
+ * Sat  5 Apr 16:35:14 UTC 2025
+ */
+
+extern void setup_LCD();
+
+/* end. */
